requirements.txt
django
djangorestframework
django-cors-headers
Pillow
django-filter
